Update test_guess_the_word.py
I played around with 3 different tests, but left two because the second was not helpful at all. I realize though that I had nothing for failing gracefully in terms of the user inputting more than one letter, a number, or nothing at all. I'll have to update the main function to fix that.