Update guess_the_word.py
Updated to just add capitals for the feedback given to the user so it looks nicer. Next I'm actually gonna fix some of the problems with the user input taking more than 2 letters, numbers, and empty spaces. After that if I have time I'll make replaying/rerunning the function better.