Update guess_the_word.py
Fully completed. Uppercases were implemented along with barriers for users inputting wrong values for the guesses. Everything is also wrapped in a while True statement to allow a prompt to replay the game or to close the program at the end. I could've done a better job with the if statement on the last part though since it only looks out for "no", so you could technically enter 1 and that will replay the game.